Add Fulcro config (#30)
* Initial support for Fulcro

Only few namespaces are exposed so far, but it is already
possible to render a trivial defsc component to a dom element.

FIXME: Tests currently don't run due to problems with accessing `js/ReactDOMServer`

* Dev support - shadow-powered dev repl

Add :dev profile to deps, and shadow-cljs config to build the code and serve a page with it. The entrypoint is `dev/development.cljs`. See Readme.

This makes it easy to interactively develop and test SCI configs.

* Replace ana/error with a fn, not to bloat output size